--- eips/doc.go
@@ -1,6 +1,0 @@
-// Copyright Tharsis Labs Ltd.(Evmos)
-// SPDX-License-Identifier:ENCL-1.0(https://github.com/evmos/evmos/blob/main/LICENSE)
-//
-// eips contains Evmos Improvement Proposal that will be merged with Ethereum Improment Porposal
-// to modify opcodes.
-package eips
